Fix redirecting with the "?next" parameter
The 'url' parameter of the CAS /logout URI is no longer valid in CAS 3.0.
Here's an excerpt from the docs:

"""
The url parameter defined in the former CAS 2.0 specification is not a
valid parameter in CAS 3.0 anymore. CAS Servers MUST ignore given url
parameters.
"""

Also, `get_service_url` ignored the URL passed on to it with the GETs
"next" parameter, so that it was lost when the browser was redirected back
to the login page after the successful CAS authentication. I believe that
the best fix for this is to use the `get_redirect_url` function whenever
`redirect_to` is not given explicitly. (This will also store the original
HTTP "REFERER" header for later use, which is also a good thing.)