Make Single Sign Out and Proxy Granting Ticket agnostic from session backend
We add an admin command called django_cas_ng_clean_sessions which remove old
SessionTicket and old ProxyGrantingTicket whose session are expired.
It should be called just after the command clearsessions which delete expired
sessions.
cf https://docs.djangoproject.com/en/1.8/topics/http/sessions/#clearing-the-session-store

For the admin command to be detected by django, the package must not be distributed
as an egg, that's why the zip_safe=False added in setup.py