Update README with Django 2 URL conf, headings
This makes the configuration section of the README clearer.  Previously, the requirement for adding URL mappings and syncing the database was buried down after all the optional settings.  They could be found before, but this adds headings and an introductory sentence explaining what the overall steps are so someone knows what to expect.

Also adds URL dispatcher details for the new Django 2.0 URL routing syntax, providing both old and new syntax