Remove transitive dependency lxml from requirements.txt (#371)
As part of our ongoing research on Python dependency management we
noticed a potential improvement in your project’s dependency management.

Specifically, the transitive dependency `lxml` is specified as a
requirement in the `requirements.txt` file, when in reality it is not
needed.

This PR removes it from `requirements.txt` to let pip manage it
automatically, which helps keeping the dependency list clean.

Hope this is helpful!

Best regards